feat: add widget page
